moved floats back to float32

+++ poker_solver/advantage_memory.py
@@ -17,7 +17,7 @@ class AdvantageMemory:
         # Preallocate storage
         self.cards = torch.zeros((capacity, num_cards), dtype=torch.int64, device=self.device)
         self.bets = torch.zeros((capacity, num_bets), dtype=torch.int64, device=self.device)
-        self.advantages = torch.zeros((capacity, num_actions), dtype=torch.float64, device=self.device)
+        self.advantages = torch.zeros((capacity, num_actions), dtype=torch.float32, device=self.device)
         self.index = 0
         self.full = False
